requirements.txt
pyfluidsynth==1.3.2
python-rtmidi==1.5.8
PyYAML==6.0
Flask==2.3.2
watchdog==3.0.0
sf2utils
